Refactor model files.
