+++ HealthOneWebServer/HealthOneWebServer.csproj
@@ -11,4 +11,8 @@
     <PackageReference Include="Swashbuckle.AspNetCore" Version="6.4.0" />
   </ItemGroup>
 
+  <ItemGroup>
+    <Folder Include="Model\ExerciseDbApi\Base\" />
+  </ItemGroup>
+
 </Project>
