--- HealthOneWebServer/Model/ExerciseDbApi/Base/BaseApiResponse.cs
@@ -1,9 +1,0 @@
-namespace HealthOneWebServer.Model.ExerciseDbApi.Base
-{
-  public class BaseApiResponse
-  {
-    public bool Success { get; set; }
-    public object? Metadata { get; set; }
-
-  }
-}
