+++ HealthOneWebServer/Model/ExerciseDbApi/Dto/ExerciseEntity.cs
@@ -1,6 +1,6 @@
 namespace HealthOneWebServer.Model.ExerciseDbApi.Exercise
 {
-  public class ExerciseDto
+  public class ExerciseEntity
   {
     public string ExerciseId { get; set; }
     public string Name { get; set; }
@@ -9,16 +9,7 @@ namespace HealthOneWebServer.Model.ExerciseDbApi.Exercise
     public List<string> Instructions { get; set; }
     public List<string> BodyParts { get; set; }
     public List<string> Equipments { get; set; }
-    public string GifUrl { get; set; }
+    public string? GifUrl { get; set; }
 
-  }
-
-  public class MetaData
-  {
-    public int CurrentPage { get; set; }
-    public string NextPage { get; set; }
-    public string PreviousPage { get; set; }
-    public int TotalExercises { get; set; }
-    public int TotalPages { get; set; }
   }
 }
