--- HealthOneWebServer/Model/ExerciseDbApi/Exercise/SingleExerciseResponseDto.cs
@@ -1,9 +1,0 @@
-using HealthOneWebServer.Model.ExerciseDbApi.Base;
-
-namespace HealthOneWebServer.Model.ExerciseDbApi.Exercise
-{
-  public class SingleExerciseResponseDto : BaseApiResponse
-  {
-    public ExerciseDto? Data { get; set; }
-  }
-}
